backend/requirements.txt
boto3
requests

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

